requirements.txt
torch
transformers
sentence-transformers
langchain
langchain_community
langchain-huggingface
langchain_experimental
langchain_chroma
streamlit
langchain_community
langchain-google-genai
python-dotenv
streamlit
langchain_experimental
sentence-transformers
langchain_chroma
langchainhub
pypdf
rapidocr-onnxruntime